porting rdkit-dependent logic to RDKitToolkitWrapper
